python_app/requirements.txt
Flask==2.2.3
flask-cors==4.0.0
Werkzeug==2.2.3
numpy==1.24.2
pandas==1.5.3
mysql-connector-python==8.0.33
openpyxl==3.1.2